gh/workflows: Add datapath conformance suite
This commit introduces the GH action "ci-datapath". The action is used
to run the cilium-cli's datapath connectivity suite from [1]. Currently,
it is running on a single VM with a single configuration as an early
proof whether the "ci-datapath" pipeline is going to be sound.

The new action introduces two novelties when compared to the existing GH
actions in Cilium.

Firstly, we are using the "self-hosted" runner. This is handled by
[2]. In short, for each instance of the action, the [2] project creates
a VM which will run the action steps.

Secondly, we are using the "brb/lvh-action" [3] (TODO: to move to either
isovalent/ or cilium/) which creates a LVH-based lightweight VM [4] on
which the test cmds are being run.

The next steps are the following:
- Keep extending the datapath connectivity suite [5]
- Create a workflow matrix for different configurations and kernels, and
  introduce some parallelism (the "self-hosted" runner VM is capable of
  running multiple instances of LVH VMs at the same time).

[1]: https://github.com/cilium/cilium-cli/blob/master/connectivity/suite.go
[2]: https://github.com/isovalent/poor-mans-ephemeral-github-action-runner
[3]: https://github.com/brb/lvh-action
[4]: https://github.com/cilium/little-vm-helper
[5]: https://github.com/cilium/cilium/issues/20606

Signed-off-by: Martynas Pumputis <m@lambda.lt>